requirements.txt
openai>=1.40.0
chromadb>=0.4.22
python-dotenv>=1.0.0
tiktoken>=0.6.0
streamlit>=1.28.0
